# Slight optimization of fallback for is_ascii